Strip out redundant code
